Rename class and files Damage CL
